Add Mediapipe Metadata handling
Only basic test
