+++ src/BUILD
@@ -200,6 +200,8 @@ cc_library(
         "mediapipe_internal/calculator_graph.hpp",
         "mediapipe_internal/mediapipedemo.cpp",
         "mediapipe_internal/mediapipedemo.hpp",
+        "mediapipe_internal/mediapipegraphdefinition.cpp",
+        "mediapipe_internal/mediapipegraphdefinition.hpp",
         "mediapipe_internal/ovms_calculator.cc",
         "metric.cpp",
         "metric.hpp",
